pull melody gen example into a header
